Only start Kafka/ZK if taskmanager.distributed is true. (#1112)

+++ grakn-dist/src/bin/grakn.sh
@@ -46,21 +46,23 @@ clean_grakn() {
     fi
 
     if cd "${GRAKN_HOME}/logs"; then
-        rm -f *.log
-        echo "Deleted logs in `pwd`" >&2
-        cd - >/dev/null
+      rm -f *.log
+      echo "Deleted logs in `pwd`" >&2
+      cd - >/dev/null
     fi
 
-    local DEFAULT_KAFKA_LOGS=/tmp/grakn-kafka-logs/
-    if [ -e "${DEFAULT_KAFKA_LOGS}" ]; then
+    if [ $USE_KAFKA ]; then
+      local DEFAULT_KAFKA_LOGS=/tmp/grakn-kafka-logs/
+      if [ -e "${DEFAULT_KAFKA_LOGS}" ]; then
         rm -rf "${DEFAULT_KAFKA_LOGS}"
         echo "Deleted logs in ${DEFAULT_KAFKA_LOGS}" >&2
-    fi
+      fi
 
-    local DEFAULT_ZOOKEEPER_LOGS=/tmp/grakn-zookeeper/
-    if [ -e "${DEFAULT_ZOOKEEPER_LOGS}" ]; then
+      local DEFAULT_ZOOKEEPER_LOGS=/tmp/grakn-zookeeper/
+      if [ -e "${DEFAULT_ZOOKEEPER_LOGS}" ]; then
         rm -rf "${DEFAULT_ZOOKEEPER_LOGS}"
         echo "Deleted logs in ${DEFAULT_ZOOKEEPER_LOGS}" >&2
+      fi
     fi
 }
 
@@ -73,16 +75,20 @@ start)
     if [ $USE_CASSANDRA ]; then
         "${GRAKN_HOME}/bin/grakn-cassandra.sh" start
     fi
-    "${GRAKN_HOME}/bin/zookeeper-server-start.sh" -daemon "${GRAKN_HOME}/conf/kafka/zookeeper.properties"
-    "${GRAKN_HOME}/bin/kafka-server-start.sh" -daemon "${GRAKN_HOME}/conf/kafka/kafka.properties"
+    if [ $USE_KAFKA ]; then
+      "${GRAKN_HOME}/bin/zookeeper-server-start.sh" -daemon "${GRAKN_HOME}/conf/kafka/zookeeper.properties"
+      "${GRAKN_HOME}/bin/kafka-server-start.sh" -daemon "${GRAKN_HOME}/conf/kafka/kafka.properties"
+    fi
     "${GRAKN_HOME}/bin/grakn-engine.sh" start
     ;;
 
 stop)
 
     "${GRAKN_HOME}/bin/grakn-engine.sh" stop
-    "${GRAKN_HOME}/bin/kafka-server-stop.sh"
-    "${GRAKN_HOME}/bin/zookeeper-server-stop.sh"
+    if [ $USE_KAFKA ]; then
+      "${GRAKN_HOME}/bin/kafka-server-stop.sh"
+      "${GRAKN_HOME}/bin/zookeeper-server-stop.sh"
+    fi
     if [ $USE_CASSANDRA ]; then
         "${GRAKN_HOME}/bin/grakn-cassandra.sh" stop
     fi
@@ -91,8 +97,10 @@ stop)
 clean)
 
     "${GRAKN_HOME}/bin/grakn-engine.sh" stop
-    "${GRAKN_HOME}/bin/kafka-server-stop.sh"
-    "${GRAKN_HOME}/bin/zookeeper-server-stop.sh"
+    if [ $USE_KAFKA ]; then
+      "${GRAKN_HOME}/bin/kafka-server-stop.sh"
+      "${GRAKN_HOME}/bin/zookeeper-server-stop.sh"
+    fi
     if [ $USE_CASSANDRA ]; then
         "${GRAKN_HOME}/bin/grakn-cassandra.sh" stop
         clean_grakn
@@ -102,8 +110,10 @@ clean)
 status)
 
     "${GRAKN_HOME}/bin/grakn-engine.sh" status
-    "${GRAKN_HOME}/bin/kafka-server-status.sh"
-    "${GRAKN_HOME}/bin/zookeeper-server-status.sh"
+    if [ $USE_KAFKA ]; then
+      "${GRAKN_HOME}/bin/kafka-server-status.sh"
+      "${GRAKN_HOME}/bin/zookeeper-server-status.sh"
+    fi
     if [ $USE_CASSANDRA ]; then
         "${GRAKN_HOME}/bin/grakn-cassandra.sh" status
     fi
